Move generator code for specific database objects (so far only tables) out into separate class
