+++ DynamicLinqPadPostgreSqlDriver/IDatabaseObjectProvider.cs
@@ -1,0 +1,10 @@
+﻿using System.Reflection.Emit;
+using LINQPad.Extensibility.DataContext;
+
+namespace DynamicLinqPadPostgreSqlDriver
+{
+    public interface IDatabaseObjectProvider
+    {
+        ExplorerItem EmitCodeAndGetExplorerItemTree(TypeBuilder dataContextTypeBuilder);
+    }
+}
